Move code and implement #12
